Feature/internal methods validation (#74)
* Add CAIP10 validation

* Add session request and notification validation

* Final clean up

* Add WC uri validation. Add permissions validation to session approve

* Final clean up

* Add peer errors responses

* Review corrections

* Add missing notification authorization check

* Fix the optional meta data params deserialization

* Review corrections

+++ walletconnectv2/src/main/kotlin/com/walletconnect/walletconnectv2/core/model/vo/JsonRpcResponseVO.kt
@@ -20,7 +20,7 @@ internal sealed class JsonRpcResponseVO {
     ) : JsonRpcResponseVO()
 
     internal data class Error(
-        val code: Long,
+        val code: Int,
         val message: String,
     )
 }
